+++ common/src/main/java/io/github/flemmli97/flan/gui/FakePlayerScreenHandler.java
@@ -1,6 +1,5 @@
 package io.github.flemmli97.flan.gui;
 
-import io.github.flemmli97.flan.api.permission.PermissionRegistry;
 import io.github.flemmli97.flan.claim.Claim;
 import io.github.flemmli97.flan.claim.PermHelper;
 import io.github.flemmli97.flan.config.ConfigHandler;
@@ -18,28 +17,28 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 
 import java.util.List;
+import java.util.UUID;
 
-public class GroupScreenHandler extends ServerOnlyScreenHandler<Claim> {
+public class FakePlayerScreenHandler extends ServerOnlyScreenHandler<Claim> {
 
     private final Claim claim;
-
     private boolean removeMode;
 
-    private GroupScreenHandler(int syncId, Inventory playerInventory, Claim claim) {
+    private FakePlayerScreenHandler(int syncId, Inventory playerInventory, Claim claim) {
         super(syncId, playerInventory, 6, claim);
         this.claim = claim;
     }
 
-    public static void openGroupMenu(Player player, Claim claim) {
+    public static void open(Player player, Claim claim) {
         MenuProvider fac = new MenuProvider() {
             @Override
             public AbstractContainerMenu createMenu(int syncId, Inventory inv, Player player) {
-                return new GroupScreenHandler(syncId, inv, claim);
+                return new FakePlayerScreenHandler(syncId, inv, claim);
             }
 
             @Override
             public Component getDisplayName() {
-                return PermHelper.simpleColoredText(ConfigHandler.langManager.get("screenGroups"));
+                return PermHelper.simpleColoredText(ConfigHandler.langManager.get("screenMenuFakePlayers"));
             }
         };
         player.openMenu(fac);
@@ -47,11 +46,12 @@ public class GroupScreenHandler extends ServerOnlyScreenHandler<Claim> {
 
     @Override
     protected void fillInventoryWith(Player player, SeparateInv inv, Claim claim) {
+        List<String> players = claim.getAllowedFakePlayerUUID();
         for (int i = 0; i < 54; i++) {
             if (i == 0) {
-                ItemStack stack = new ItemStack(Items.TNT);
-                stack.setHoverName(ServerScreenHelper.coloredGuiText(ConfigHandler.langManager.get("screenBack"), ChatFormatting.DARK_RED));
-                inv.updateStack(i, stack);
+                ItemStack close = new ItemStack(Items.TNT);
+                close.setHoverName(ServerScreenHelper.coloredGuiText(ConfigHandler.langManager.get("screenBack"), ChatFormatting.DARK_RED));
+                inv.updateStack(i, close);
             } else if (i == 3) {
                 ItemStack stack = new ItemStack(Items.ANVIL);
                 stack.setHoverName(ServerScreenHelper.coloredGuiText(ConfigHandler.langManager.get("screenAdd"), ChatFormatting.DARK_GREEN));
@@ -63,13 +63,12 @@ public class GroupScreenHandler extends ServerOnlyScreenHandler<Claim> {
             } else if (i < 9 || i > 44 || i % 9 == 0 || i % 9 == 8)
                 inv.updateStack(i, ServerScreenHelper.emptyFiller());
             else {
-                List<String> groups = claim.groups();
                 int row = i / 9 - 1;
                 int id = (i % 9) + row * 7 - 1;
-                if (id < groups.size()) {
-                    ItemStack group = new ItemStack(Items.PAPER);
-                    group.setHoverName(ServerScreenHelper.coloredGuiText(String.format(ConfigHandler.langManager.get("screenGroupName"), groups.get(id)), ChatFormatting.DARK_BLUE));
-                    inv.updateStack(i, group);
+                if (id < players.size()) {
+                    ItemStack fakePlayer = new ItemStack(Items.ZOMBIE_HEAD);
+                    fakePlayer.setHoverName(ServerScreenHelper.coloredGuiText(String.format(ConfigHandler.langManager.get("screenFakePlayerNameUUID"), players.get(id)), ChatFormatting.YELLOW));
+                    inv.updateStack(i, fakePlayer);
                 }
             }
         }
@@ -91,13 +90,18 @@ public class GroupScreenHandler extends ServerOnlyScreenHandler<Claim> {
         if (index == 3) {
             player.closeContainer();
             player.getServer().execute(() -> StringResultScreenHandler.createNewStringResult(player, (s) -> {
-                this.claim.editPerms(player, s, PermissionRegistry.EDITPERMS, -1);
+                boolean fl = player.getServer().getProfileCache().get(s).map(prof -> this.claim.modifyFakePlayerUUID(prof.getId(), false)).orElse(true);
                 player.closeContainer();
-                player.getServer().execute(() -> GroupScreenHandler.openGroupMenu(player, this.claim));
-                ServerScreenHelper.playSongToPlayer(player, SoundEvents.ANVIL_USE, 1, 1f);
+                player.getServer().execute(() -> FakePlayerScreenHandler.open(player, this.claim));
+                if (fl)
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.ANVIL_USE, 1, 1f);
+                else {
+                    player.displayClientMessage(PermHelper.simpleColoredText(ConfigHandler.langManager.get("playerGroupAddFail"), ChatFormatting.RED), false);
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.VILLAGER_NO, 1, 1f);
+                }
             }, () -> {
                 player.closeContainer();
-                player.getServer().execute(() -> GroupScreenHandler.openGroupMenu(player, this.claim));
+                player.getServer().execute(() -> FakePlayerScreenHandler.open(player, this.claim));
                 ServerScreenHelper.playSongToPlayer(player, SoundEvents.VILLAGER_NO, 1, 1f);
             }));
             ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
@@ -113,20 +117,15 @@ public class GroupScreenHandler extends ServerOnlyScreenHandler<Claim> {
         }
         ItemStack stack = slot.getItem();
         if (!stack.isEmpty()) {
-            String name = stack.getHoverName().getString();
-            if (this.removeMode) {
-                this.claim.removePermGroup(player, name);
+            UUID uuid = null;
+            try {
+                uuid = UUID.fromString(stack.getHoverName().getString());
+            } catch (IllegalArgumentException ignored) {
+            }
+            if (this.removeMode && uuid != null) {
+                this.claim.modifyFakePlayerUUID(uuid, true);
                 slot.set(ItemStack.EMPTY);
                 ServerScreenHelper.playSongToPlayer(player, SoundEvents.BAT_DEATH, 1, 1f);
-            } else {
-                if (clickType == 1) {
-                    player.closeContainer();
-                    player.getServer().execute(() -> PermissionScreenHandler.openClaimMenu(player, this.claim, name));
-                } else {
-                    player.closeContainer();
-                    player.getServer().execute(() -> GroupPlayerScreenHandler.openPlayerGroupMenu(player, this.claim, name));
-                }
-                ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
             }
         }
         return false;